requirements.txt
Adafruit_DHT
Adafruit-SSD1306
Pillow
smbus
